backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0